fix: Default version of SSL/TLS
To fix the problem, we should replace the deprecated ssl.wrap_socket method with a more secure and modern approach. The recommended way is to use ssl.SSLContext or ssl.create_default_context to ensure that a secure protocol like TLS 1.2 or above is used.

Create an SSLContext object and set its minimum_version to ssl.TLSVersion.TLSv1_2.
Wrap the server socket using this context instead of ssl.wrap_socket.